+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## master
 
+## 0.8.1 (2024-03-27)
+
 - Added `channel.whisper(...)`. ([@palkan][])
 
   Clients can send transient publications to channels via whispering. NOTE: it must be supported by the server and enabled for the channel. See [docs](https://docs.anycable.io/edge/anycable-go/signed_streams#whispering).